+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **[BREAKING CHANGE]** Import `icons` as default instead of named export for a better code splitting
 - **[UPDATE]** Make `Review` more semantic with proper HTML and Review/Rating microdata
 - **[FIX]** Fix clicking oustide of `SearchForm` closes overlays
+- **[UPDATE]** Use `Divider` in `ItemList` instead of custom CSS
   [...]
 
 # v25.0.0 (02/04/2020)