+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - **[UPDATE]** Make `Review` more semantic with proper HTML and Review/Rating microdata
 - **[FIX]** Fix clicking oustide of `SearchForm` closes overlays
 - **[UPDATE]** Use `Divider` in `ItemList` instead of custom CSS
+- **[UPDATE]** Remove legacy style from Autocomplete
   [...]
 
 # v25.0.0 (02/04/2020)